* transport.js: Resolved an intermediate problem under IE where the connection would fail because the inline script block used to create the IFrame's content would not appear to run soon enough before it needed to be used.
The IFrame source now references about:blank as its content, which is an immediate operation under IE6->9 and allows the connection to succeed as the IFrame is immediately ready for use.

Attempting to reference a blank.html file instead of about:blank is not desirable either, because on the initial page load (without any cache) the IFrame takes too long to load (from testing, ~15-30ms) and by then the connection was already trying to be established by using the IFrame.

Subsequent page loads did work however when the blank.html file was served from cache (~1ms load time) as the load was quick enough.

By using about:blank as the IFrame source, every page load (cached or not) always has the IFrame ready for use and the connection attempt always succeeds when possible.