bandaid to remove leftover initial slashes on relative paths on windows
Some (but not all) import paths in files that get compiled for the
simulator on windows are built with a mix of a correct '\' separated
directory path and a '/' separated path at the end relative to the
devkit root directory. During the import process, this gets resolved
into a relative path by simply removing all the matching path at the
front compared to the directory, which leads to a path starting with /
that jsio then mistakes as an absolute path, which breaks the import in
the compile stage and incorrectly has the browser attempt to load the
file itself.

This is a bandaid over the problem to make the simulator work on
windows. A full fix should fix the relative path construction in a cross
platform way, fix the way relative vs absolute paths are detected in a
cross platform way, and fix the way relative paths are resolved (can we use
node's path.resolve function in all cases?). This is non-trivial to fix
without breaking all the different ways things are imported through jsio
(browser vs node, *nix vs windows, relative vs absolute, etc).