fix (makeRelativePath): strip \ and / when resolving paths
After removing a subpath when resolving a relative path,
look for a path separator (/ or \, regardless of platform), and
remove it if necessary. This prevents mixed paths (eg:
c:\initial\folder/added/path) from being resolved with a leftover / on
the front on systems with \ as the path separator (windows).